nav: add Dashboard + Watchlist links; let logged-in users see /
Three changes around the logged-in/logged-out nav:

- Drop the `/` → `/account` redirect for signed-in visitors. The
  marketing homepage is now reachable for everyone by clicking the
  logo. Login still lands on /account by default (auth callback's
  `next` defaults to `/account`).
- Mark `/` as `force-dynamic` to replace the dynamic rendering hint
  we lost when removing the getUser() call. Keeps live data off the
  build-time static HTML.
- Add two new top-nav links — "Dashboard" (/account) and
  "Watchlist" (/account/watchlist) — visible only when signed in,
  slotted in front of "Leaderboard". Session state is now resolved
  inside <Nav /> itself (was inside <NavAuth /> alone) so the link
  set can depend on it; NavAuth is refactored to receive email +
  ready as props rather than running its own listener.
